feat(memories): add Memory interface and implement new components for displaying memories
