+++ types.d.ts
@@ -1,0 +1,10 @@
+
+export interface Memory {
+  id: number;
+  src: string;
+  caption: string;
+  shortTitle: string;
+  category: string;
+  year: string;
+  color: string;
+}
